Issue #7277 - Using HttpConfiguration for configuration
